pyproject.toml
[tool.poetry]
name = "cartoes_crescer"
version = "0.1.0"
description = ""
authors = ["HM"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
django = "^4.2.15"
djangorestframework = "3.15.2"  
django-cors-headers = "4.4.0"   
django-filter = "^23.2"
django-environ = "0.10.0"
django-countries="7.5.1"
openpyxl="3.1.5"
phonenumbers="8.13.7"
argon2-cffi = "^23.1.0"
django-allauth = "0.55.2"
djangorestframework-simplejwt = "5.2.2"
dj-rest-auth = "5.0.1"
drf-yasg= "1.21.7" 
psycopg2-binary = "^2.9.9"
pandas="2.2.3"
django-autoslug = "^1.9.9"
pillow = "^10.4.0"
gunicorn = "20.1.0"



[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
black=  "^23.1.0"
isort = "^5.12.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
[tool.black]
line-length = 88
target-version = ['py311']
extend-exclude = '''
(
  /(
    \.eggs
    | \.git
    | \.mypy_cache
    | \.envs
    | \.tox
    | \.venv
    | \env
    | \conda_env
    | conda_env
    | env
    | venv
    | build
  )/
)
'''

[tool.isort]
profile = "black"

[tool.flake8]
ignore = "E203, E501, W503"
max-line-length = 120